requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
streamlit-authenticator==0.2.3
pyyaml>=6.0
numpy>=1.24.0
pandas>=1.5.0
altair>=4.0 